Move workspace creation into its own file.
