+++ src/entities/Folder.ts
@@ -1,5 +1,5 @@
 /* eslint-disable import/prefer-default-export */
-import { Entity, ManyToMany, Property, Collection, Unique } from '@mikro-orm/core'
+import { Entity, ManyToMany, Property, Collection } from '@mikro-orm/core'
 
 // eslint-disable-next-line import/no-cycle
 import { Workspace } from './index';
@@ -9,11 +9,7 @@ import { BaseEntity } from './BaseEntity';
 export class Folder extends BaseEntity
 {
 
-  // @PrimaryKey({ type: 'number' })
-  // id!: number;
-
   @Property({ type: 'string' })
-  @Unique()
   path!: string;
 
   @ManyToMany(() => Workspace, workspace => workspace.folders)
